requirements.txt
Django>=4.2
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-cloud-vision>=3.7.0
